apps/adjudication_engine/requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpickle==3.1.2
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
Faker==40.8.0
fastapi==0.135.1
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-genai==1.66.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
llvmlite==0.46.0
MarkupSafe==3.0.3
motor==3.7.1
narwhals==2.18.0
numba==0.64.0
numpy==2.4.3
ollama==0.6.1
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pillow==12.1.1
plotly==6.6.0
pluggy==1.6.0
prometheus_client==0.24.1
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pymongo==4.16.0
PyMuPDF==1.27.2
pyparsing==3.3.2
pytesseract==0.3.13
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
pytz==2026.1.post1
referencing==0.37.0
reportlab==4.4.10
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
shap==0.51.0
six==1.17.0
slicer==0.0.8
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.52.1
streamlit==1.55.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
watchdog==6.0.0
websockets==16.0
xgboost==3.2.0
